requirements.txt
asgiref==3.7.2
attrs==22.1.0
autocomplete==0.0.104
backports.zoneinfo==0.2.1
bottle==0.12.23
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cms==0.0.2
cryptography==38.0.1
dal-admin-filters==1.1.0
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.1.1
django-admin-interface==0.22.1
django-allauth==0.51.0
django-autocomplete-light==3.9.4
django-colorfield==0.7.2
django-cors-headers==4.3.1
django-environ==0.9.0
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-markdown==0.8.4
django-nvd3==0.9.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
environ==1.0
factory-boy==3.2.1
Faker==14.2.0
idna==3.4
importlib-metadata==4.12.0
iniconfig==1.1.1
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.3
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
pipreqs==0.4.11
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
types-cryptography==3.3.23
typing_extensions==4.4.0
Unidecode==1.3.7
urllib3==1.26.12
yarg==0.1.9
zipp==3.8.1
